Passing all tests again after refactoring.
