+++ grade/pipeline/__init__.py
@@ -1,2 +1,5 @@
 from .asserts import *
+from .partialcredit import PartialCredit
 from .pipeline import *
+from .run import Run
+from .write import WriteOutputs, WriteStderr, WriteStdout
